Move auth state to services
