+++ lib/features/profile/screens/profile.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
+import '../../../services/auth_state.dart';
 import '../../../utils/extensions.dart';
-import '../../auth/providers/auth_state.dart';
 import '../providers/profile.dart';
 
 class ProfileScreen extends ConsumerWidget {
