+++ lib/services/auth_state.dart
@@ -1,9 +1,8 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../../services/api/api_service.dart';
-import '../../../services/storage/secure_storage.dart';
-import '../models/auth_state.dart';
-import '../models/login.dart';
+import '../features/auth/models/login.dart';
+import 'api/api_service.dart';
+import 'storage/secure_storage.dart';
 
 part 'auth_state.g.dart';
 
@@ -51,4 +50,49 @@ class CurrentAuthState extends _$CurrentAuthState {
       // previous token.
       ..invalidate(tokenProvider);
   }
+}
+
+/// The possible authentication states of the app.
+enum AuthState {
+  unknown(
+    redirectPath: '/',
+    allowedPaths: [
+      '/',
+    ],
+  ),
+  unauthenticated(
+    redirectPath: '/login',
+    allowedPaths: [
+      '/login',
+      '/settings',
+    ],
+  ),
+  authenticated(
+    redirectPath: '/products',
+    allowedPaths: [
+      '/products',
+      '/products/:id',
+      '/todos',
+      '/todos/add',
+      '/todos/:id',
+      '/todos/:id/update',
+      '/posts',
+      '/posts/:id',
+      '/profile',
+      '/settings',
+    ],
+  ),
+  ;
+
+  const AuthState({
+    required this.redirectPath,
+    required this.allowedPaths,
+  });
+
+  /// The target path to redirect when the current route is not allowed in this
+  /// auth state.
+  final String redirectPath;
+
+  /// List of paths allowed when the app is in this auth state.
+  final List<String> allowedPaths;
 }
