+++ lib/services/router.dart
@@ -2,8 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../features/auth/models/auth_state.dart';
-import '../features/auth/providers/auth_state.dart';
 import '../features/auth/screens/login.dart';
 import '../features/posts/screens/post.dart';
 import '../features/posts/screens/posts.dart';
@@ -16,6 +14,7 @@ import '../features/todos/screens/todo.dart';
 import '../features/todos/screens/todos.dart';
 import '../features/todos/screens/update_todo.dart';
 import '../widgets/scaffold_with_navigation.dart';
+import 'auth_state.dart';
 
 part 'router.g.dart';
