+++ test/features/profile/providers/profile_test.dart
@@ -1,8 +1,8 @@
 import 'package:dummymart/features/auth/models/login.dart';
-import 'package:dummymart/features/auth/providers/auth_state.dart';
 import 'package:dummymart/features/profile/models/profile.dart';
 import 'package:dummymart/features/profile/providers/profile.dart';
 import 'package:dummymart/services/api/api_service.dart';
+import 'package:dummymart/services/auth_state.dart';
 import 'package:dummymart/services/storage/prefs.dart';
 import 'package:dummymart/services/storage/secure_storage.dart';
 import 'package:test/test.dart';
